[Diem] - switched to ALPHA2 - made easier to override front actions and components base class - allowed php command in admin console - improved error handlers on admin diagrams - added warnings in mail utilities - added £media default configuration - many improvments in installer. No more need to launch dm:setup task after installation. - moved the doctrine forms unsetAutoFields() call to base classes - added project setup check in dmContext->initialize - made dmLogEntry show localhost ip - made PluginDmMedia->checkFileExists() more robust - removed deprecated dmFilesystem->whois method - added graphviz recommandation to dmServerCheck, and removed deprecated ones - fixed file backup permissions - fixed usage of mbstring in dmSeoSynchronizer - improved dmSetupTask - removed deprecated dmUpgradeTask->upgradeTo500Dev1 method - improved browser server check stylesheet - fixed front code editor error handler - added dmValidatorCssClasses - added help messages on front text widget - fixed warning when generating front templates - improved generated front templates comments - added min and max values to dmValidatorCssSize - added save button and confirmation messages to front code editor - fixed static calls to i18n service in front widget forms - improved secure page

+++ dmCorePlugin/lib/model/doctrine/PluginDmMedia.class.php
@@ -115,7 +115,7 @@ abstract class PluginDmMedia extends BaseDmMedia
 
     $exists = file_exists($this->getFullPath());
 
-    if (false === $exists && $orDelete)
+    if (false === $exists && $orDelete && $this->exists())
     {
       $this->delete();
     }
